fix: drop use of injectables
