--- crates/macros/src/common/inject.rs
@@ -1,5 +1,0 @@
-use proc_macro::TokenStream;
-
-pub(crate) fn inject_macro(_args: TokenStream, raw_input: TokenStream) -> TokenStream {
-    raw_input
-}
