--- crates/macros/src/common/mod.rs
@@ -1,5 +1,0 @@
-pub mod check;
-pub mod handler;
-pub mod inject;
-pub mod injectable;
-pub mod state;
