--- crates/macros/src/core/mod.rs
@@ -1,3 +1,0 @@
-pub mod dto;
-pub mod param;
-pub mod query;
