--- crates/macros/src/utils/str.rs
@@ -1,14 +1,0 @@
-use syn::Ident;
-
-/// Converts a string into an Ident.
-///
-/// #### Arguments
-///
-/// * `string` - The string to convert into an Ident.
-///
-/// #### Returns
-///
-/// The Ident representation of the string.
-pub(crate) fn str_to_ident(string: String) -> Ident {
-    Ident::new(&string, proc_macro::Span::call_site().into())
-}
