+++ crates/shared/src/traits/mod.rs
@@ -1,9 +1,7 @@
 mod gate_trait;
-mod injectable_trait;
 mod interpreter;
 mod middleware_trait;
 
 pub use gate_trait::*;
-pub use injectable_trait::*;
 pub use interpreter::*;
 pub use middleware_trait::*;
